requirements.txt
pandas
numpy
requests
python-dotenv
scikit-learn
matplotlib
plotly
streamlit